requirements.txt
aiofiles==22.1.0
aioredis==2.0.1
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.9.14
click==8.0.3
h11==0.12.0
httpcore==0.15.0
httptools==0.5.0
httpx==0.23.0
idna==3.4
jsonschema==3.2.0
multidict==5.2.0
PyJWT==2.3.0
pymongo==3.11.4
pyrsistent==0.18.1
python-dotenv==0.19.2
PyYAML==6.0
rfc3986==1.5.0
sanic==21.6.1
Sanic-Cors==1.0.1
sanic-openapi==21.6.1
sanic-plugin-toolkit==1.2.1
sanic-redis==0.3.0
sanic-routing==0.7.2
sanic-session==0.8.0
sanic-testing==22.6.0
six==1.16.0
sniffio==1.3.0
termcolor==1.1.0
typing-extensions==4.3.0
ujson==5.5.0
uvloop==0.17.0
websockets==10.3
